build: add new controller and Thymeleaf to use html

+++ src/main/java/com/dev/myninjaregister/ninjaregister1/Missions/MissionsController.java
@@ -1,5 +1,9 @@
 package com.dev.myninjaregister.ninjaregister1.Missions;
 
+import com.dev.myninjaregister.ninjaregister1.Common.Response.GenericAnswer;
+import org.apache.catalina.connector.Response;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,26 +19,82 @@ public class MissionsController {
     }
 
     @PostMapping("/create")
-    public MissionsDTO createMission(@RequestBody MissionsDTO mission){
-        return missionsService.createNewMission(mission);
+    public ResponseEntity<String> createMission(@RequestBody MissionsDTO mission) {
+        MissionsDTO missionsDTO = missionsService.createNewMission(mission);
+        return ResponseEntity.status(HttpStatus.OK)
+                .body("The mission was created!");
     }
+
     @GetMapping("/showall")
-    public List<MissionsDTO> showAllMissions(){
-        return missionsService.showAllMissions();
+    public ResponseEntity<GenericAnswer<List<MissionsDTO>>> showAllMissions() {
+        List<MissionsDTO> missionsList = missionsService.showAllMissions();
+        if (missionsList != null) {
+            GenericAnswer<List<MissionsDTO>> answer = new GenericAnswer<>(
+                    missionsList,
+                    "This s the full list of missions"
+
+            );
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(answer);
+        } else {
+            GenericAnswer<List<MissionsDTO>> noAnswer = new GenericAnswer<>(
+                    null,
+                    "Do no exist any mission"
+            );
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(noAnswer);
+        }
     }
+
     @GetMapping("/showbyid/{id}")
-    public MissionsDTO showMissionById(@PathVariable Long id){
-        return missionsService.showMissionsById(id);
+    public ResponseEntity<GenericAnswer<MissionsDTO>> showMissionById(@PathVariable Long id) {
+        MissionsDTO missionsDTOid = missionsService.showMissionsById(id);
+        if (missionsDTOid != null) {
+            GenericAnswer<MissionsDTO> answer = new GenericAnswer<>(
+                    missionsDTOid,
+                    "The mission was found"
+            );
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(answer);
+        } else {
+            GenericAnswer<MissionsDTO> noAnswer = new GenericAnswer<>(
+                    null,
+                    "The mission with this id was not found"
+            );
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(noAnswer);
+        }
     }
 
     @PutMapping("/editbyid/{id}")
-    public MissionsDTO updateMission(@PathVariable Long id, @RequestBody MissionsDTO attmission){
-        return missionsService.modifyMission(id,attmission);
+    public ResponseEntity<GenericAnswer<MissionsDTO>> updateMission(@PathVariable Long id, @RequestBody MissionsDTO attmission) {
+        MissionsDTO missionsDTOedit = missionsService.modifyMission(id, attmission);
+        if (missionsDTOedit != null) {
+            GenericAnswer<MissionsDTO> answer = new GenericAnswer<>(
+                    missionsDTOedit,
+                    "The ninjas was edited"
+            );
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(answer);
+        } else {
+            GenericAnswer<MissionsDTO> noAnswer = new GenericAnswer<>(
+                    null,
+                    "The mission was not found"
+            );
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(noAnswer);
+        }
     }
 
     @DeleteMapping("/deletebyid/{id}")
-    public void deleteMission(@PathVariable Long id){
-        missionsService.deleteMissionById(id);
+    public ResponseEntity<String> deleteMission(@PathVariable Long id) {
+        if (missionsService.showMissionsById(id) != null) {
+            missionsService.deleteMissionById(id);
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body("The ninja was deleted");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("The ninjas was not found");
+        }
     }
-
 }
